pyproject.toml
[project]
name = "mcp-airflow-documentation"
version = "0.1.0"
description = "MCP server for Apache Airflow documentation search"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0.0,<3.0.0",
    "docutils>=0.21.0",
    "python-frontmatter>=1.1.0",
    "click>=8.0.0",
]

[project.scripts]
mcp-airflow-documentation = "mcp_airflow_documentation.server:main"
airflow-docs-index = "mcp_airflow_documentation.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=src/mcp_airflow_documentation --cov-report=term-missing --cov-report=html"
